+++ src/Ext/Aircraft/Hooks.cpp
@@ -3,17 +3,23 @@
 
 
 #include <Ext/Aircraft/Body.h>
+#include <Ext/Techno/Body.h>
 #include <Ext/Anim/Body.h>
 #include <Ext/WeaponType/Body.h>
 #include <Utilities/Macro.h>
+
+#pragma region Strafing
 
 DEFINE_HOOK(0x417FF1, AircraftClass_Mission_Attack_StrafeShots, 0x6)
 {
 	GET(AircraftClass*, pThis, ESI);
 
+	auto pExt = TechnoExt::ExtMap.Find(pThis);
+
 	if (pThis->MissionStatus < (int)AirAttackStatus::FireAtTarget2_Strafe
 		|| pThis->MissionStatus >(int)AirAttackStatus::FireAtTarget5_Strafe)
 	{
+		pExt->Strafe_BombsDroppedThisRound = 0;
 		return 0;
 	}
 
@@ -21,13 +27,56 @@ DEFINE_HOOK(0x417FF1, AircraftClass_Mission_Attack_StrafeShots, 0x6)
 	auto pWeaponExt = WeaponTypeExt::ExtMap.Find(pThis->GetWeapon(weaponIndex)->WeaponType);
 	int fireCount = pThis->MissionStatus - 4;
 
-	if (fireCount > 1 && pWeaponExt->Strafing_Shots < fireCount)
+	if (pWeaponExt->Strafing_Shots > 5)
+	{
+		if (pThis->MissionStatus == (int)AirAttackStatus::FireAtTarget3_Strafe)
+		{
+			int remainingShots = pWeaponExt->Strafing_Shots - 3 - pExt->Strafe_BombsDroppedThisRound;
+
+			if (remainingShots > 0)
+				pThis->MissionStatus = (int)AirAttackStatus::FireAtTarget2_Strafe;
+		}
+	}
+	else if (fireCount > 1 && pWeaponExt->Strafing_Shots < fireCount)
 	{
 		if (!pThis->Ammo)
-			pThis->unknown_bool_6D2 = false;
+			pThis->IsLocked = false;
 
 		pThis->MissionStatus = (int)AirAttackStatus::ReturnToBase;
 	}
+
+	return 0;
+}
+
+__forceinline bool AircraftCanStrafeWithWeapon(WeaponTypeClass const* pWeapon)
+{
+	return pWeapon && WeaponTypeExt::ExtMap.Find(pWeapon)->Strafing
+		.Get(pWeapon->Projectile->ROT <= 1 && !pWeapon->Projectile->Inviso);
+}
+
+long __stdcall AircraftClass_IFlyControl_IsStrafe(IFlyControl const* ifly)
+{
+	__assume(ifly != nullptr);
+	auto pThis = static_cast<AircraftClass const*>(ifly);
+
+	if (!pThis->Target || pThis->Target->IsInAir())
+		return false;
+
+	int weaponIndex = pThis->SelectWeapon(pThis->Target);
+	return (long)AircraftCanStrafeWithWeapon(pThis->GetWeapon(weaponIndex)->WeaponType);
+}
+
+DEFINE_JUMP(VTABLE, 0x7E2268, GET_OFFSET(AircraftClass_IFlyControl_IsStrafe));
+
+DEFINE_HOOK(0x415F25, AircraftClass_Fire, 0x6)
+{
+	GET(AircraftClass*, pThis, EDI);
+	GET(BulletClass*, pTraj, ESI);
+
+	__assume(pThis != nullptr);
+
+	if (AircraftCanStrafeWithWeapon(pTraj->WeaponType))
+		TechnoExt::ExtMap.Find(pThis)->Strafe_BombsDroppedThisRound++;
 
 	return 0;
 }
@@ -36,7 +85,7 @@ DEFINE_HOOK(0x418403, AircraftClass_Mission_Attack_FireAtTarget_BurstFix, 0x8)
 {
 	GET(AircraftClass*, pThis, ESI);
 
-	pThis->unknown_bool_6C8 = true;
+	pThis->ShouldLoseAmmo = true;
 
 	AircraftExt::FireBurst(pThis, pThis->Target, 0);
 
@@ -87,6 +136,8 @@ DEFINE_HOOK(0x418B1F, AircraftClass_Mission_Attack_FireAtTarget5Strafe_BurstFix,
 
 	return 0x418B40;
 }
+
+#pragma endregion
 
 DEFINE_HOOK(0x414F10, AircraftClass_AI_Trailer, 0x5)
 {